Dendrite project/requirements.txt
Flask==2.1.4
Flask-Security-Too==3.3.0
Flask-GraphQL-Auth==0.1.2
Flask-Stripe==0.5.0
graphene==2.17.3
python-dotenv==0.19.1
Flask-CORS==3.1.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
